user-service/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
pymongo==4.3.3
bcrypt==4.0.1
PyJWT==2.6.0